Index1024.jl

Documentation for Index1024.jl

Index1024.build_index_file — Function

build_index_file(io::IO, kvs; meta=String[], aux=Dict())
build_index_file(filename::AbstractString, kvs; meta=String[], aux=Dict())

#Arguments io::IO descriptor for writing (so you can use IOBuffer if desired) meta::Vector{AbstractString} vector of strings to add meta data Create the on-disk representation of the index of the kvs Dict. The Tree's Leaves are sorted by the key value of the kvs and store both the kvs[key] and auxkey All keys and values are all converted to UInt64.

Index1024.open_index — Function

open_index(filename::AbstractString)::Index
open_index(io::IO)::Index

Create an Index struct on which one can perform searches using a previously created Index file.

Index1024.search — Function

search(idx::Index, search_key::UInt64)::Union{Tuple{UInt64, UInt64}, Nothing}

Search the given index for a given searchkey returning a Tuple of the previously stored value / aux pair (or 0 for the aux if it wasn't supplied). If the searchkey is not found, return nothing

Base.get — Function

get(idx::Index, search_key, default)

Search the given index for a given searchkey returning a Tuple of the previously stored value / aux pair (or 0 for the aux if it wasn't supplied). If the searchkey is not found, return the supplied default